pyproject.toml
[project]
name = "mobiclaw"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # --- Core agent framework (workspace member) ---
    "agentscope",
    # --- Web server ---
    "fastapi",
    "uvicorn[standard]",
    "python-multipart",   # required for FastAPI UploadFile
    "pydantic",           # directly imported in gateway_server.py
    # --- Scheduling ---
    "apscheduler",
    # --- HTTP client ---
    "requests",
    # --- Messaging integrations ---
    "lark-oapi>=1.4.0",   # Feishu/Lark SDK (lazy-loaded in gateway_server.py)
    # --- Mobile agent (vision) ---
    "pillow",             # PIL.Image used in mobile providers
    # --- Document processing (lazy-imported in tools/) ---
    "pypdf<=6.5.0",       # papers.py; version cap matches agentscope[pdf-reader]
    "openpyxl",           # office.py
    "reportlab",
    "python-pptx>=0.6.23",
    "python-docx",        # office.py  (imported as 'docx')
    # --- RAG / vector store ---
    "qdrant-client==1.15.1",  # agentscope.rag.QdrantStore; pinned to avoid pymilvus conflict
    "nltk",               # agentscope text/PDF/Word readers
    # --- Markdown conversion (CLI tool used in skill_runner tests) ---
    "markitdown",
    # --- Testing ---
    "pytest",
    "jsonschema",         # test_agents_tool_json_schemas.py
    "uiautomator2>=3.5.0",
    "hmdriver2>=1.4.4",
]

[tool.uv.workspace]
members = [
    "agentscope",
]

[tool.uv.sources]
agentscope = { workspace = true }
